+++ src/__tests__/actions/FETCH_DATA.js
@@ -1,0 +1,9 @@
+import FETCH_DATA from '../../actions/FETCH_DATA';
+
+jest.mock('../../actions/FETCH_DATA');
+
+describe('FETCH_DATA', () => {
+  it('should return an object', () => {
+    expect(FETCH_DATA()).toReturn(typeof Function);
+  });
+});
