+++ src/actions/FETCH_DATA.js
@@ -2,11 +2,6 @@ import axios from 'axios';
 
 const baseUrl = `https://api.nomics.com/v1/currencies/ticker?key=${process.env.REACT_APP_API_KEY}&per-page=25&interval=1d&convert=USD&sort=rank`;
 
-const LOADING = data => ({
-  type: 'LOADING',
-  payload: data,
-});
-
 const FETCH_DATA = () => {
   const request = axios.get(baseUrl);
   return dispatch => {
@@ -19,9 +14,4 @@ const FETCH_DATA = () => {
   };
 };
 
-const FILTER = currency => ({
-  type: 'FILTER',
-  payload: currency,
-});
-
-export { LOADING, FETCH_DATA, FILTER };
+export default FETCH_DATA;
