+++ src/actions/FILTER.js
@@ -1,0 +1,6 @@
+const FILTER = currency => ({
+  type: 'FILTER',
+  payload: currency,
+});
+
+export default FILTER;
