+++ src/containers/App.js
@@ -10,7 +10,8 @@ import Header from '../components/Header';
 import FeaturedComponent from '../components/FeaturedComponent';
 import Attribution from '../components/Attribution';
 import CurrenciesContainer from './CurrenciesContainer';
-import { FETCH_DATA, FILTER } from '../actions/index';
+import FILTER from '../actions/FILTER';
+import FETCH_DATA from '../actions/FETCH_DATA';
 import FilterCurrency from '../components/FilterCurrency';
 import Details from '../components/Details';
 import HomePage from '../skeleton/HomePage';
